Fix courses image, Add Skeleton
