+++ src/pages/Payment/Payment.jsx
@@ -1,0 +1,7 @@
+import React, { useState } from 'react'
+
+function Payment() {
+  return <></>
+}
+
+export default Payment
