--- src/pages/Payment/index.jsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function index() {
-  return <div>this is payment page</div>
-}
